Add Command-, Message-, and InteractionHandlers
