+++ src/types/needleCommand.ts
@@ -1,0 +1,7 @@
+import { SlashCommandBuilder } from "@discordjs/builders";
+import { BaseCommandInteraction } from "discord.js";
+
+export interface NeedleCommand {
+	info: SlashCommandBuilder;
+	execute(interaction: BaseCommandInteraction): Promise<void>;
+}
